feat:  DB 수정 반영해서 코드 리뉴얼

+++ src/main/java/net/skhu/repository/DepartmentRepository.java
@@ -1,9 +1,7 @@
 package net.skhu.repository;
 
+import net.skhu.domain.Department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import net.skhu.domain.Department;
-
-public interface DepartmentRepository extends JpaRepository<Department, Integer> {
-	
+public interface DepartmentRepository extends JpaRepository<Department, Long> {
 }
